refactor: split attention surface helpers
